pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "luau-bench"
version = "0.0.1"
description = "An evaluation harness for benchmarking LLM-generated Luau code."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Luau Bench Contributors"},
]
keywords = ["benchmark", "luau", "llm", "evaluation", "harness", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "rich>=13.0",
    "click>=8.1",
    "toml>=0.10",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "tqdm>=4.65",
]
progress = [
    "tqdm>=4.65",
]
anthropic = [
    "anthropic>=0.25",
]
openai = [
    "openai>=1.0",
]
datasets = [
    "datasets>=2.14",
]
all = [
    "openai>=1.0",
    "anthropic>=0.25",
    "datasets>=2.14",
    "tqdm>=4.65",
]

[project.scripts]
luau-bench = "luau_bench.cli:main"

[tool.setuptools.packages.find]
include = ["luau_bench*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
